+++ src/main/java/odyssey/backend/directory/dto/response/DirectoryResponse.java
@@ -1,9 +1,9 @@
-package odyssey.backend.directory.dto;
+package odyssey.backend.directory.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import odyssey.backend.directory.domain.Directory;
-import odyssey.backend.roadmap.dto.SimpleRoadmapResponse;
+import odyssey.backend.roadmap.dto.response.SimpleRoadmapResponse;
 
 import java.util.List;
